voice-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
speechmatics-batch==2.0.1
thymia-sentinel==1.1.0
